requirements.txt
google-generativeai>=0.7.0
streamlit>=1.34
PyYAML>=6.0